Full Refactor initial commit
