+++ .gitignore
@@ -1,0 +1,23 @@
+# Byte-compiled / optimized / DLL files
+__pycache__/
+*.py[cod]
+*$py.class
+
+# Distribution / packaging
+*.egg-info/
+*.egg
+
+# Installer logs
+pip-log.txt
+pip-delete-this-directory.txt
+
+#Jupyter
+.ipynb_checkpoints
+
+# Environments
+.env
+.venv
+
+# IDE
+.vscode/
+
