+++ errant/__init__.py
@@ -1,0 +1,2 @@
+from errant.errant import Errant
+from errant.version import VERSION as __version__
