+++ errant/version.py
@@ -1,0 +1,6 @@
+_MAJOR = '0'
+_MINOR = '1'
+_REVISION = '0'
+
+VERSION_SHORT = '{0}.{1}'.format(_MAJOR, _MINOR)
+VERSION = '{0}.{1}.{2}'.format(_MAJOR, _MINOR, _REVISION)
